+++ tests/Doubles/TestMaintenanceChecker.php
@@ -1,0 +1,15 @@
+<?php declare(strict_types = 1);
+
+namespace Tests\OriNette\Scheduler\Doubles;
+
+use Orisai\Scheduler\Maintenance\MaintenanceChecker;
+
+final class TestMaintenanceChecker implements MaintenanceChecker
+{
+
+	public function isMaintenance(): bool
+	{
+		return false;
+	}
+
+}